Updating SoS chests to work closer to OSI.
Players should be able to have a chance to get a Arty, and a decor /
tmap.  This chance would be super rare, but it should be possible.  Also
moved the creation of Deco items outside of the bag as they appeared in
the base chest along with the maps that hsould have a possibility of
appearing.  Also removed deco items from arty items as they should be
seperate and not take away from getting an arty.