Fixes part 2 of closed #649
Fixes dropping an amount from a stack and the bounce back to the
original stack (for all stacked items, not just scrolls). This
intentionally does not use Serialization, any items caught in that state
will fall through to the old default of dropping back into its parent
container or world location.