Update Virulent.cs
Virulent is named mobile, and should also be moved to that directory. Should the corpse names of these named versions of basic creature mobiles be 'an X corpse' or 'the remains of X'? It's not listed on stratics or uoguide, and I've seen both. 'an X corpse' is more common, so I've used that here.