Basic logging microframework with colored console logging support.
Introduces a basic framework for logging, with interface similar to log4net, which provides typical log levels.
At the moment, the provided implementation just outputs the messages to stdout, with a color depending on the log level.
This is good to standarize levels of message outputs and to decouple client code from log streams and settings, such as log category, formatting, etc.