Update RunicAtlas.cs
The runebook will tell you if you try to recall but you do not have the recall spell.

The Runic Atlas was not doing that. It would just do nothing leaving the user confused, until they realized they did not have a sellbook with that spell on them. I have fixed it here.