Update BaseClothing.cs
Since LowerAmmoCost was added as a basic property it should appear in here too.

Currently if you were to put LowerAmmoCost on a piece of clothing it would not appear. Even though there are no LowerAmmoCost clothing items, yet, there are also no LowerAmmoCost weapons yet it correctly appears on those.