Update Kraken.cs
Currently on ServUO krakens spawn in the open ocean. They should not.

On top of that the Kraken that is "fished up" is the only one the drops MIBS. The few that spawn in dungeons 100% do not.

Since SpecialFishingNet.cs already determines the chance for a fished up one to get a MIB there is no need for the repeat method on the Kraken itself. On top of that the "check" for its location is also not needed since the ones that spawn in dungeons do not drop MIBS and the current ServUO repo is wrong with having them spawn in the wild, open oceans, without having to be fished up with fishing nets.

Tested and confirmed on EA.