Update BaseBoat.cs
All credit goes to Lokai for this one! 

Under the old decay system it NEVER said like new because it was always >= Also this way allows shard owners to easily set decay times and not have to edit this method either. Tested with timers of 2 minutes and 10 minutes and it works great.