Stop from defaulting to 90 min cooldown (#5045)
It looks like the code was missing an else-if statement. The current servuo code will always set the duration to 30 if the skill is greater than or equal to 120, and then set it to 60 if the skill is greater than or equal to 100, and THEN set it to 90 if 60 regardless of whether the duration has already been set to 30. This will properly set the timer.